requirements/dev.txt
-r base.txt

coverage~=3.7
django-debug-toolbar~=1.7
Fabric~=1.13.2
factory-boy~=2.8.1
ipython~=3.0.0
mock~=1.0.1
pytest-django~=3.1.2
pytest~=3.0.7
